% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_energy_in}
\alias{parse_energy_in}
\title{Parse Energy In}
\usage{
parse_energy_in(name, cat)
}
\arguments{
\item{name}{name from energypro}

\item{cat}{category from energypro}
}
\description{
Parse Energy In
}
