% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_energy_out}
\alias{parse_energy_out}
\title{Get Energy Out}
\usage{
parse_energy_out(name, cat)
}
\arguments{
\item{name}{name from energypro}

\item{cat}{category from energypro}
}
\description{
Get Energy Out
}
