% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_energypro.R
\name{read_energypro}
\alias{read_energypro}
\title{Read Energypro}
\usage{
read_energypro(path)
}
\arguments{
\item{path}{path to csv file with energypro raw data}
}
\description{
Read energypro data from a csv file and return a tidy tibble
}
