% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conversion.R
\name{unit_conversion}
\alias{unit_conversion}
\title{Unit Conversion}
\usage{
unit_conversion(unit_actual, unit_target)
}
\arguments{
\item{unit_actual}{actual unit}

\item{unit_target}{target unit}
}
\description{
Unit Conversion
}
